Split the form into separate files. Fix the loading state.
