+++ apps/studio/data/table-editor/table-editor-query.ts
@@ -1,10 +1,11 @@
-import { QueryClient, useQuery } from '@tanstack/react-query'
+import { QueryClient, queryOptions, useQuery } from '@tanstack/react-query'
+import { id } from 'common-tags'
+import { UseCustomQueryOptions } from 'types'
 
 import { executeSql, ExecuteSqlError } from '../sql/execute-sql-query'
 import { tableEditorKeys } from './keys'
 import { getTableEditorSql } from './table-editor-query-sql'
 import { Entity } from './table-editor-types'
-import { UseCustomQueryOptions } from 'types'
 
 type TableEditorArgs = {
   id?: number
@@ -48,8 +49,7 @@ export const useTableEditorQuery = <TData = TableEditorData>(
   }: UseCustomQueryOptions<TableEditorData, TableEditorError, TData> = {}
 ) =>
   useQuery<TableEditorData, TableEditorError, TData>({
-    queryKey: tableEditorKeys.tableEditor(projectRef, id),
-    queryFn: ({ signal }) => getTableEditor({ projectRef, connectionString, id }, signal),
+    ...tableEditorQueryOptions({ projectRef, connectionString, id }),
     enabled:
       enabled && typeof projectRef !== 'undefined' && typeof id !== 'undefined' && !isNaN(id),
     refetchOnWindowFocus: false,
@@ -62,16 +62,16 @@ export function prefetchTableEditor(
   client: QueryClient,
   { projectRef, connectionString, id }: TableEditorVariables
 ) {
-  return client.fetchQuery({
+  return client.fetchQuery(tableEditorQueryOptions({ projectRef, connectionString, id }))
+}
+
+export const tableEditorQueryOptions = <TData = TableEditorData>({
+  projectRef,
+  connectionString,
+  id,
+}: TableEditorVariables) => {
+  return queryOptions<TableEditorData, TableEditorError, TData>({
     queryKey: tableEditorKeys.tableEditor(projectRef, id),
-    queryFn: ({ signal }) =>
-      getTableEditor(
-        {
-          projectRef,
-          connectionString,
-          id,
-        },
-        signal
-      ),
+    queryFn: ({ signal }) => getTableEditor({ projectRef, connectionString, id }, signal),
   })
 }
